Restructure for maintenance and collaboration. Added a classic joke
